+++ docs/make.jl
@@ -2,6 +2,7 @@ cd(@__DIR__)
 using Attractors
 using Attractors.DynamicalSystemsBase
 using Attractors.StateSpaceSets
+using PredefinedDynamicalSystems
 
 pages = [
     "index.md",